requirements.txt
crewai==0.41.1
crewai-tools==0.4.26
langchain-openai==0.1.17
langchain-google-genai==1.0.10
PyPDF2==3.0.1
pdfplumber==0.11.0
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings
colorama==0.4.6
rich==13.7.1
